# docs/swaggerui/api_spec.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "PHP / Symfony example Tasks REST API",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8080/",
            "description": "Development Server"
        }
    ],
    "paths": {
        "/v1/tasks/{taskId}": {
            "get": {
                "description": "get Task by ID",
                "operationId": "",
                "parameters": [
                    {
                        "name": "taskId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Task"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "unauthorized"
                    },
                    "404": {
                        "description": "task not found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "put": {
                "description": "update Task by ID",
                "operationId": "App\\Controller\\TasksController::updateTask",
                "parameters": [
                    {
                        "name": "taskId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "duedate",
                                    "completed"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "My Task"
                                    },
                                    "duedate": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2020-01-01"
                                    },
                                    "completed": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Task"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "missing title"
                    },
                    "401": {
                        "description": "unauthorized"
                    },
                    "404": {
                        "description": "task not found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "delete": {
                "description": "delete Task by ID",
                "operationId": "App\\Controller\\TasksController::deleteTask",
                "parameters": [
                    {
                        "name": "taskId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "success"
                    },
                    "401": {
                        "description": "unauthorized"
                    },
                    "404": {
                        "description": "task not found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/v1/tasks": {
            "get": {
                "description": "get (un-)completed Tasks",
                "operationId": "App\\Controller\\TasksController::getTasks",
                "parameters": [
                    {
                        "name": "completed",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Task"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "unauthorized"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "post": {
                "description": "create new Task",
                "operationId": "App\\Controller\\TasksController::createTask",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "duedate"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "My Task"
                                    },
                                    "duedate": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2020-01-01"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Task"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "missing title"
                    },
                    "401": {
                        "description": "unauthorized"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "Task": {
                "properties": {
                    "id": {
                        "description": "ID",
                        "type": "integer",
                        "example": 42
                    },
                    "title": {
                        "description": "Title",
                        "type": "string",
                        "example": "My Task"
                    },
                    "duedate": {
                        "description": "Due date",
                        "type": "string",
                        "format": "date",
                        "example": "2020-01-01"
                    },
                    "completed": {
                        "description": "Completed",
                        "type": "boolean",
                        "example": false
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearer": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}